chore: moved some of the main code to different files
Not really a big commit or anything but this one just made the code in
the main loop a bit more seperated.

I'm also planning on rewriting the pool allocator because it does really
serve me well right now.

This isn't a problem of the implementation perse but rather that its a
stack pool rather than a heap pool.

No worries I found a paper online that goes into depth on pool
allocators :)